requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.1
python-multipart==0.0.9
pydantic==1.10.14
passlib==1.7.4
argon2-cffi==23.1.0
python-jose==3.3.0
requests==2.31.0
dnspython==2.6.1
bcrypt==4.1.2
